+++ src/main/java/wooteco/subway/dto/LineResponse.java
@@ -1,0 +1,34 @@
+package wooteco.subway.dto;
+
+import wooteco.subway.domain.Line;
+
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+
+    public LineResponse() {
+    }
+
+    public LineResponse(Long id, String name, String color) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public static LineResponse from(Line line) {
+        return new LineResponse(line.getId(), line.getName(), line.getColor());
+    }
+}
