--- src/main/java/wooteco/subway/line/LineControllerAdvice.java
@@ -1,25 +1,0 @@
-package wooteco.subway.line;
-
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class LineControllerAdvice {
-
-    @ExceptionHandler(EmptyResultDataAccessException.class)
-    public ResponseEntity<String> handleEmptyResultException(Exception e) {
-        return ResponseEntity
-                .badRequest()
-                .body(e.getMessage());
-    }
-
-    @ExceptionHandler(DuplicateKeyException.class)
-    public ResponseEntity<String> handleDuplicateUniqueColumnException(Exception e) {
-        return ResponseEntity
-                .badRequest()
-                .body("이미 존재하는 노선 이름입니다.");
-    }
-}
