--- src/main/java/wooteco/subway/station/Section.java
@@ -1,61 +1,0 @@
-package wooteco.subway.station;
-
-import java.util.Objects;
-
-public class Section {
-    private final Station upStation;
-    private final Station downStation;
-    private final int distance;
-
-    public Section(Station upStation, Station downStation, int distance) {
-        this.upStation = upStation;
-        this.downStation = downStation;
-        this.distance = distance;
-    }
-
-    public boolean hasAny(Station station) {
-        return this.upStation.equals(station) || this.downStation.equals(station);
-    }
-
-    public boolean isUpStation(Station station) {
-        return this.upStation.equals(station);
-    }
-
-    public boolean isDownStation(Station station) {
-        return this.downStation.equals(station);
-    }
-
-    public int getDistance() {
-        return this.distance;
-    }
-
-    public Station getUpStation() {
-        return this.upStation;
-    }
-
-    public Station getDownStation() {
-        return this.downStation;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Section section = (Section) o;
-        return distance == section.distance && Objects.equals(upStation, section.upStation) && Objects.equals(downStation, section.downStation);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(upStation, downStation, distance);
-    }
-
-    @Override
-    public String toString() {
-        return "Section{" +
-                "upStation=" + upStation +
-                ", downStation=" + downStation +
-                ", distance=" + distance +
-                '}';
-    }
-}
