--- src/main/java/wooteco/subway/station/StationControllerAdvice.java
@@ -1,25 +1,0 @@
-package wooteco.subway.station;
-
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class StationControllerAdvice {
-
-    @ExceptionHandler(EmptyResultDataAccessException.class)
-    public ResponseEntity<String> handleEmptyResultException(Exception e) {
-        return ResponseEntity
-                .badRequest()
-                .body(e.getMessage());
-    }
-
-    @ExceptionHandler(DuplicateKeyException.class)
-    public ResponseEntity<String> handleDuplicateUniqueColumnException(Exception e) {
-        return ResponseEntity
-                .badRequest()
-                .body("이미 존재하는 역 이름입니다.");
-    }
-}
